pyproject.toml
[project]
name = "fundingscape"
version = "0.1.0"
description = "EU Research Funding Landscape Intelligence System"
requires-python = ">=3.12"

dependencies = [
    "duckdb>=1.0",
    "httpx>=0.27",
    "hishel>=0.0.31",
    "pydantic>=2.7",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "click>=8.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-httpx>=0.30",
    "pytest-asyncio>=0.23",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/fundingscape"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
